PLT-294 Added slash command to check used version of CSH (#28)
* PLT-294 Added slash command to check used version of CSH

"require version csh" is useful in .csh scripts (init or not),
where it can tell CSH to: exit, error or warn when a version constraint fails.

* PLT-294 Added wildcard support to "require version csh"

So it's now possible to require version ==2.5.*

* PLT-294 Rename error-level to error-action and fatal to quit

* PLT-294 Added unit tests to require version

They are failing of course. So I guess I have to fix that.

* PLT-294 Made parse_version() a lot more flexible and fixed operator for compare_version()

Also expanded the unit tests quite a bit.

* PLT-294 Forgot to push meson.build update

* PLT-294 Some more tests and fixes for compare_version()

* PLT-294 Forgot to commit header file

* PLT-294 Forgot to compile new file dedicated to the "require version csh" slash command.

* PLT-294 Allow semver suffix, so we can parse CSH version itself.

We also no longer use a mutable copy of the version string,
so we allow much larger versions.